Reorganization, with apologies
