+++ site/src/api/index.test.ts
@@ -1,5 +1,6 @@
 import axios from "axios"
-import { APIKeyResponse, getApiKey, login, LoginResponse, logout } from "./api"
+import { getApiKey, login, logout } from "."
+import { LoginResponse, APIKeyResponse } from "./types"
 
 // Mock the axios module so that no real network requests are made, but rather
 // we swap in a resolved or rejected value
