+++ site/src/pages/projects/[organization]/[project]/create.tsx
@@ -3,6 +3,7 @@ import { makeStyles } from "@material-ui/core/styles"
 import { useNavigate, useParams } from "react-router-dom"
 import useSWR from "swr"
 
+import * as Types from "../../../../api/types"
 import * as API from "../../../../api"
 import { ErrorSummary } from "../../../../components/ErrorSummary"
 import { FullScreenLoader } from "../../../../components/Loader/FullScreenLoader"
@@ -14,11 +15,11 @@ export const CreateWorkspacePage: React.FC = () => {
   const navigate = useNavigate()
   const styles = useStyles()
 
-  const { data: organizationInfo, error: organizationError } = useSWR<API.Organization, Error>(
+  const { data: organizationInfo, error: organizationError } = useSWR<Types.Organization, Error>(
     () => `/api/v2/users/me/organizations/${organizationName}`,
   )
 
-  const { data: project, error: projectError } = useSWR<API.Project, Error>(() => {
+  const { data: project, error: projectError } = useSWR<Types.Project, Error>(() => {
     return `/api/v2/organizations/${unsafeSWRArgument(organizationInfo).id}/projects/${projectName}`
   })
 
@@ -26,7 +27,7 @@ export const CreateWorkspacePage: React.FC = () => {
     navigate(`/projects/${organizationName}/${projectName}`)
   }, [navigate, organizationName, projectName])
 
-  const onSubmit = async (req: API.CreateWorkspaceRequest) => {
+  const onSubmit = async (req: Types.CreateWorkspaceRequest) => {
     const workspace = await API.Workspace.create(req)
     navigate(`/workspaces/${workspace.id}`)
     return workspace
