+++ site/src/pages/workspaces/[workspace].tsx
@@ -8,24 +8,24 @@ import { ErrorSummary } from "../../components/ErrorSummary"
 import { FullScreenLoader } from "../../components/Loader/FullScreenLoader"
 import { Workspace } from "../../components/Workspace"
 import { unsafeSWRArgument } from "../../util"
-import * as API from "../../api"
+import * as Types from "../../api/types"
 
 export const WorkspacePage: React.FC = () => {
   const styles = useStyles()
   const { workspace: workspaceQueryParam } = useParams()
 
-  const { data: workspace, error: workspaceError } = useSWR<API.Workspace, Error>(() => {
+  const { data: workspace, error: workspaceError } = useSWR<Types.Workspace, Error>(() => {
     const workspaceParam = firstOrItem(workspaceQueryParam, null)
 
     return `/api/v2/workspaces/${workspaceParam}`
   })
 
   // Fetch parent project
-  const { data: project, error: projectError } = useSWR<API.Project, Error>(() => {
+  const { data: project, error: projectError } = useSWR<Types.Project, Error>(() => {
     return `/api/v2/projects/${unsafeSWRArgument(workspace).project_id}`
   })
 
-  const { data: organization, error: organizationError } = useSWR<API.Project, Error>(() => {
+  const { data: organization, error: organizationError } = useSWR<Types.Project, Error>(() => {
     return `/api/v2/organizations/${unsafeSWRArgument(project).organization_id}`
   })
